Refactor quest and achievement reward handling to be more centralized
- Create a new RewardHelper class that contains the majority of shared reward handling
- Move a lot of methods from QuestRewardHelper into RewardHelper
- Fix a bug in `applyMoneyBoost` that could result in compounding money boost being applied across characters
- Route achievement reward handling through RewardHelper
