+++ project/assets/database/locales/server/en.json
@@ -594,9 +594,7 @@
     "quest-handover_wrong_item": "Unable to hand item in for quest: {{questId}}, expected tpl: {{requiredTpl}} but handed in: {{handedInTpl}}",
     "quest-item_not_found_in_inventory": "changeItemStack() Item with _id: %s not found in inventory",
     "quest-no_skill_found": "Skill %s not found",
-    "quest-reward_type_not_handled": "Quest reward type: {{rewardType}} not handled for quest: {{questId}} name: {{questName}}",
     "quest-unable_to_find_compare_condition": "Unrecognised Comparison Method: %s",
-    "quest-unable_to_find_matching_hideout_production": "Unable to find matching hideout craft unlock for quest: {{questName}}, matches found: {{matchCount}}",
     "quest-unable_to_find_quest_in_db": "Quest id: {{questId}} with type: {{questType}} not found in database",
     "quest-unable_to_find_quest_in_db_no_quest_rewards": "Unable to find quest: %s in db, unable to give quest rewards to player",
     "quest-unable_to_find_repeatable_to_replace": "Unable to find repeatable quest in profile to replace, skipping",
@@ -651,6 +649,8 @@
     "repeatable-quest_handover_failed_condition_invalid": "Quest handover error: condition not found or incorrect value. qid: {{body.qid}}, condition: {{body.conditionId}}",
     "repeatable-unable_to_accept_quest_see_log": "Unable to accept quest, see server log for details",
     "repeatable-unable_to_accept_quest_starting_message_not_found": "Unable to accept quest: {{questId}} cant find quest started message text with id: {{messageId}}",
+    "reward-type_not_handled": "Reward type: {{rewardType}} not handled for quest/achievement: {{questId}}",
+    "reward-unable_to_find_matching_hideout_production": "Unable to find matching hideout craft unlock for quest/achievement: {{questId}}, matches found: {{matchCount}}",
     "route_onupdate_no_response": "onUpdate: %s route doesn't report success or fail",
     "scav-missing_karma_level_getting_default": "getScavKarmaLevel() failed, unable to find fence in profile.traderInfo. Defaulting to karma level 0",
     "scav-missing_karma_settings": "Unable to get karma settings for level %s",
