+++ project/src/helpers/ProfileHelper.ts
@@ -547,50 +547,6 @@ export class ProfileHelper {
         return pmcProfile.Info.Bans.some((ban) => ban.banType === BanType.RAGFAIR && currentTimestamp < ban.dateTime);
     }
 
-    /**
-     * Add an achievement to player profile + check for and add any hideout customisation unlocks to profile
-     * @param fullProfile Profile to add achievement to
-     * @param achievementId Id of achievement to add
-     */
-    public addAchievementToProfile(fullProfile: ISptProfile, achievementId: string): void {
-        // Add achievement id to profile with timestamp it was unlocked
-        fullProfile.characters.pmc.Achievements[achievementId] = this.timeUtil.getTimestamp();
-
-        // Check for any customisation unlocks
-        const achievementDataDb = this.databaseService
-            .getTemplates()
-            .achievements.find((achievement) => achievement.id === achievementId);
-        if (!achievementDataDb) {
-            return;
-        }
-
-        // Get customisation reward object from achievement db
-        const customizationDirectReward = achievementDataDb.rewards.find(
-            (reward) => reward.type === "CustomizationDirect",
-        );
-        if (!customizationDirectReward) {
-            return;
-        }
-
-        const customisationDataDb = this.databaseService
-            .getHideout()
-            .customisation.globals.find((customisation) => customisation.itemId === customizationDirectReward.target);
-        if (!customisationDataDb) {
-            this.logger.error(
-                `Unable to find customisation data for ${customizationDirectReward.target} in profile ${fullProfile.info.id}`,
-            );
-
-            return;
-        }
-
-        // Reward found, add to profile
-        fullProfile.customisationUnlocks.push({
-            id: customizationDirectReward.target,
-            source: CustomisationSource.ACHIEVEMENT,
-            type: customisationDataDb.type as CustomisationType,
-        });
-    }
-
     public hasAccessToRepeatableFreeRefreshSystem(pmcProfile: IPmcData): boolean {
         return [GameEditions.EDGE_OF_DARKNESS, GameEditions.UNHEARD].includes(<any>pmcProfile.Info?.GameVersion);
     }
